backend/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-filter==25.1
django-js-asset==3.1.2
django-modeltranslation==0.19.16
django-redis==6.0.0
django-rosetta==0.10.2
django-unfold==0.62.0
djangorestframework==3.16.0
drf-spectacular==0.27.0
django-ratelimit==4.1.0
django-prometheus==2.3.1
sentry-sdk[django]==1.40.0
python-json-logger==2.0.7
django-debug-toolbar==4.2.0
django-extensions==3.2.3
celery==5.3.4
redis==6.2.0
idna==3.10
pillow==11.3.0
polib==1.2.0
psycopg2-binary==2.9.10
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
